requirements.txt
django==4.2.6
django-environ==0.11.2
openpyxl==3.1.2
pandas==2.1.1
xlwings==0.30.12
dj-database-url==2.1.0
APScheduler==3.10.4